refact(step1): move collective method to 1st class collection Answers
