+++ src/main/java/qna/domain/Answers.java
@@ -1,0 +1,21 @@
+package qna.domain;
+
+import qna.CannotDeleteException;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Answers {
+
+    private Answers(){
+
+    }
+
+    public static List<DeleteHistory> ofDelete(List<Answer> answers, User loginUser) throws CannotDeleteException {
+        List<DeleteHistory> deleteHistories = new ArrayList<>();
+        for (Answer answer : answers) {
+            deleteHistories.add(answer.delete(loginUser));
+        }
+        return deleteHistories;
+    }
+}
